refactor: slice의 actions와 reducer가 구분되도록 export와 디렉토리 구조 업데이트 (#26)
