+++ src/slices/counter/actions.ts
@@ -1,0 +1,1 @@
+export { increment, decrement, incrementByAmount } from './counterSlice';
