+++ src/slices/user/actions.ts
@@ -1,0 +1,1 @@
+export { login, logout } from './userSlice';
